Backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pymysql>=1.1.0
google-generativeai>=0.3.0
